requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-cors-headers==4.3.1
django-online-users==0.3
django-phonenumber-field==7.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environs==10.0.0
executing==2.0.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
ipykernel==6.26.0
ipython==8.17.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
phonenumbers==8.13.24
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
ratelimit==2.2.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
slugify==0.0.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.9
wheel==0.42.0
